// YOUCOM.ReserVook/src/app/features/staffs/model/staff.model.ts
/** ユーザー情報 */
export class Staff {
  displayCompanyNo: string;
  userEmail: string;
  password: string;
  userName: string;
  roleDivision: string;
  status: string;
  version :number;
  creator:string;
  updator:string;
  cdt:string;
  udt:string;

  roleName:string;
  companyName:string;
  userEmailOrigin:string;

  companyGroupId: string;
  companyNo: string;
}